requirements.txt
appdirs
asgiref
black
click
Django
flake8
gunicorn
gevent
mccabe
mypy-extensions
pathspec
pycodestyle
pyflakes
pytz
regex
sqlparse
toml
psycopg2
django-createsuperuserwithpassword
django-cors-headers
requests
whitenoise
python-decouple